Separated the usePropertiesCount function
from `Properties.tsx`.

Removed unused import from `getPropertyById.tsx`.
