+++ app/javascript/components/Property/getPropertyById.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import axios from "axios";
 import { Property } from "./useProperty";
