+++ package.json
@@ -12,6 +12,7 @@
     "@rails/actioncable": "^7.0.0",
     "@rails/activestorage": "^7.0.0",
     "@rails/ujs": "^7.0.0",
+    "@react-icons/all-files": "^4.1.0",
     "@tanstack/react-table": "^8.0.0-alpha.26",
     "@types/d3": "^7.1.0",
     "@types/history": "^4.7.9",
